Add `Storage` trait.
