+++ godot-core/src/lib.rs
@@ -50,7 +50,7 @@ pub mod private {
 
     pub use crate::gen::classes::class_macros;
     pub use crate::registry::{callbacks, ClassPlugin, ErasedRegisterFn, PluginComponent};
-    pub use crate::storage::as_storage;
+    pub use crate::storage::{as_storage, Storage};
     pub use sys::out;
 
     use crate::{log, sys};
