+++ godot-core/src/storage.rs
@@ -5,7 +5,9 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-use crate::obj::GodotClass;
+use std::ops::{Deref, DerefMut};
+
+use crate::obj::{Base, Gd, GodotClass, Inherits};
 use crate::{godot_error, out};
 use godot_ffi as sys;
 
@@ -16,312 +18,107 @@ pub enum Lifecycle {
     Destroying,
 }
 
-#[cfg(not(feature = "experimental-threads"))]
-pub(crate) use single_threaded::*;
+#[cfg_attr(feature = "experimental-threads", allow(dead_code))]
+mod single_threaded;
 
-#[cfg(feature = "experimental-threads")]
-pub(crate) use multi_threaded::*;
+#[cfg_attr(not(feature = "experimental-threads"), allow(dead_code))]
+mod multi_threaded;
 
-#[cfg(not(feature = "experimental-threads"))]
-mod single_threaded {
-    use std::any::type_name;
-    use std::cell;
+pub trait Storage {
+    type Instance: GodotClass;
+    type RefGuard<'a>: Deref<Target = Self::Instance>
+    where
+        Self: 'a;
+    type MutGuard<'a>: Deref<Target = Self::Instance> + DerefMut
+    where
+        Self: 'a;
 
-    use crate::obj::{Base, Gd, GodotClass, Inherits};
-    use crate::out;
+    fn construct(
+        user_instance: Self::Instance,
+        base: Base<<Self::Instance as GodotClass>::Base>,
+    ) -> Self;
 
-    use super::Lifecycle;
+    fn is_bound(&self) -> bool;
 
-    /// Manages storage and lifecycle of user's extension class instances.
-    pub struct InstanceStorage<T: GodotClass> {
-        user_instance: cell::RefCell<T>,
-        pub(super) base: Base<T::Base>,
+    fn base(&self) -> &Base<<Self::Instance as GodotClass>::Base>;
 
-        // Declared after `user_instance`, is dropped last
-        pub(super) lifecycle: cell::Cell<Lifecycle>,
-        godot_ref_count: cell::Cell<u32>,
+    fn get(&self) -> Self::RefGuard<'_>;
+
+    fn get_mut(&self) -> Self::MutGuard<'_>;
+
+    fn get_lifecycle(&self) -> Lifecycle;
+
+    fn set_lifecycle(&self, lifecycle: Lifecycle);
+
+    fn get_gd(&self) -> Gd<Self::Instance>
+    where
+        Self::Instance: Inherits<<Self::Instance as GodotClass>::Base>,
+    {
+        self.base().to_gd().cast()
     }
 
-    /// For all Godot extension classes
-    impl<T: GodotClass> InstanceStorage<T> {
-        pub fn construct(user_instance: T, base: Base<T::Base>) -> Self {
-            out!("    Storage::construct             <{}>", type_name::<T>());
-
-            Self {
-                user_instance: cell::RefCell::new(user_instance),
-                base,
-                lifecycle: cell::Cell::new(Lifecycle::Alive),
-                godot_ref_count: cell::Cell::new(1),
-            }
-        }
-
-        pub fn is_bound(&self) -> bool {
-            // Needs to borrow mutably, otherwise it succeeds if shared borrows are alive.
-            self.user_instance.try_borrow_mut().is_err()
-        }
-
-        pub fn get(&self) -> cell::Ref<T> {
-            self.user_instance.try_borrow().unwrap_or_else(|_e| {
-                panic!(
-                    "Gd<T>::bind() failed, already bound; T = {}.\n  \
-                     Make sure there is no &mut T live at the time.\n  \
-                     This often occurs when calling a GDScript function/signal from Rust, which then calls again Rust code.",
-                    type_name::<T>()
-                )
-            })
-        }
-
-        pub fn get_mut(&self) -> cell::RefMut<T> {
-            self.user_instance.try_borrow_mut().unwrap_or_else(|_e| {
-                panic!(
-                    "Gd<T>::bind_mut() failed, already bound; T = {}.\n  \
-                     Make sure there is no &T or &mut T live at the time.\n  \
-                     This often occurs when calling a GDScript function/signal from Rust, which then calls again Rust code.",
-                    type_name::<T>()
-                )
-            })
-        }
-
-        pub fn get_gd(&self) -> Gd<T>
-        where
-            T: Inherits<<T as GodotClass>::Base>,
-        {
-            self.base.to_gd().cast()
-        }
-
-        pub(super) fn godot_ref_count(&self) -> u32 {
-            self.godot_ref_count.get()
-        }
-
-        pub(crate) fn on_inc_ref(&self) {
-            let refc = self.godot_ref_count.get() + 1;
-            self.godot_ref_count.set(refc);
-
-            out!(
-                "    Storage::on_inc_ref (rc={})     <{}>", // -- {:?}",
-                refc,
-                type_name::<T>(),
-                //self.user_instance
-            );
-        }
-
-        pub(crate) fn on_dec_ref(&self) {
-            let refc = self.godot_ref_count.get() - 1;
-            self.godot_ref_count.set(refc);
-
-            out!(
-                "  | Storage::on_dec_ref (rc={})     <{}>", // -- {:?}",
-                refc,
-                type_name::<T>(),
-                //self.user_instance
-            );
-        }
-    }
-}
-
-#[cfg(feature = "experimental-threads")]
-mod multi_threaded {
-    use std::sync;
-    use std::sync::atomic::{AtomicU32, Ordering};
-
-    use crate::obj::{Base, Gd, GodotClass, Inherits};
-    use crate::out;
-
-    use super::Lifecycle;
-
-    pub struct AtomicLifecycle {
-        atomic: AtomicU32,
-    }
-
-    impl AtomicLifecycle {
-        pub fn new(value: Lifecycle) -> Self {
-            Self {
-                atomic: AtomicU32::new(value as u32),
-            }
-        }
-
-        pub fn get(&self) -> Lifecycle {
-            match self.atomic.load(Ordering::Relaxed) {
-                0 => Lifecycle::Alive,
-                1 => Lifecycle::Destroying,
-                other => panic!("invalid lifecycle {other}"),
-            }
-        }
-
-        pub fn set(&self, lifecycle: Lifecycle) {
-            let value = match lifecycle {
-                Lifecycle::Alive => 0,
-                Lifecycle::Destroying => 1,
-            };
-
-            self.atomic.store(value, Ordering::Relaxed);
-        }
-    }
-
-    /// Manages storage and lifecycle of user's extension class instances.
-    pub struct InstanceStorage<T: GodotClass> {
-        user_instance: sync::RwLock<T>,
-        pub(super) base: Base<T::Base>,
-
-        // Declared after `user_instance`, is dropped last
-        pub(super) lifecycle: AtomicLifecycle,
-        godot_ref_count: AtomicU32,
-    }
-
-    /// For all Godot extension classes
-    impl<T: GodotClass> InstanceStorage<T> {
-        pub fn construct(user_instance: T, base: Base<T::Base>) -> Self {
-            out!("    Storage::construct             <{:?}>", base);
-
-            Self {
-                user_instance: sync::RwLock::new(user_instance),
-                base,
-                lifecycle: AtomicLifecycle::new(Lifecycle::Alive),
-                godot_ref_count: AtomicU32::new(1),
-            }
-        }
-
-        pub fn is_bound(&self) -> bool {
-            // Needs to borrow mutably, otherwise it succeeds if shared borrows are alive.
-            self.write_ignoring_poison().is_none()
-        }
-
-        pub fn get(&self) -> sync::RwLockReadGuard<T> {
-            self.read_ignoring_poison().unwrap_or_else(|| {
-                panic!(
-                    "Gd<T>::bind() failed, already bound; obj = {}.\n  \
-                     Make sure there is no &mut T live at the time.\n  \
-                     This often occurs when calling a GDScript function/signal from Rust, which then calls again Rust code.",
-                    self.base,
-                )
-            })
-        }
-
-        pub fn get_mut(&self) -> sync::RwLockWriteGuard<T> {
-            self.write_ignoring_poison().unwrap_or_else(|| {
-                panic!(
-                    "Gd<T>::bind_mut() failed, already bound; obj = {}.\n  \
-                     Make sure there is no &T or &mut T live at the time.\n  \
-                     This often occurs when calling a GDScript function/signal from Rust, which then calls again Rust code.",
-                    self.base,
-                )
-            })
-        }
-
-        pub fn get_gd(&self) -> Gd<T>
-        where
-            T: Inherits<<T as GodotClass>::Base>,
-        {
-            self.base().clone().cast()
-        }
-
-        pub(super) fn godot_ref_count(&self) -> u32 {
-            self.godot_ref_count.load(Ordering::Relaxed)
-        }
-
-        pub(crate) fn on_inc_ref(&self) {
-            self.godot_ref_count.fetch_add(1, Ordering::Relaxed);
-            out!(
-                "    Storage::on_inc_ref (rc={})     <{:?}>",
-                self.godot_ref_count(),
-                self.base,
-            );
-        }
-
-        pub(crate) fn on_dec_ref(&self) {
-            self.godot_ref_count.fetch_sub(1, Ordering::Relaxed);
-            out!(
-                "  | Storage::on_dec_ref (rc={})     <{:?}>",
-                self.godot_ref_count(),
-                self.base,
-            );
-        }
-
-        /// Returns a write guard (even if poisoned), or `None` when the lock is held by another thread.
-        /// This might need adjustment if threads should await locks.
-        #[must_use]
-        fn write_ignoring_poison(&self) -> Option<sync::RwLockWriteGuard<T>> {
-            match self.user_instance.try_write() {
-                Ok(guard) => Some(guard),
-                Err(sync::TryLockError::Poisoned(poison_error)) => Some(poison_error.into_inner()),
-                Err(sync::TryLockError::WouldBlock) => None,
-            }
-        }
-
-        /// Returns a read guard (even if poisoned), or `None` when the lock is held by another writing thread.
-        /// This might need adjustment if threads should await locks.
-        #[must_use]
-        fn read_ignoring_poison(&self) -> Option<sync::RwLockReadGuard<T>> {
-            match self.user_instance.try_read() {
-                Ok(guard) => Some(guard),
-                Err(sync::TryLockError::Poisoned(poison_error)) => Some(poison_error.into_inner()),
-                Err(sync::TryLockError::WouldBlock) => None,
-            }
-        }
-
-        // fn __static_type_check() {
-        //     enforce_sync::<InstanceStorage<T>>();
-        // }
-    }
-
-    // TODO make InstanceStorage<T> Sync
-    // This type can be accessed concurrently from multiple threads, so it should be Sync. That implies however that T must be Sync too
-    // (and possibly Send, because with `&mut` access, a `T` can be extracted as a value using mem::take() etc.).
-    // Which again means that we need to infest half the codebase with T: Sync + Send bounds, *and* make it all conditional on
-    // `#[cfg(feature = "experimental-threads")]`.
-    //
-    // A better design would be a distinct Gds<T: Sync> pointer, which requires synchronized.
-    // This needs more design on the multi-threading front (#18).
-    //
-    // The following code + __static_type_check() above would make sure that InstanceStorage is Sync.
-    // Make sure storage is Sync in multi-threaded case, as it can be concurrently accessed through aliased Gd<T> pointers.
-    // fn enforce_sync<T: Sync>() {}
-}
-
-impl<T: GodotClass> InstanceStorage<T> {
-    pub fn debug_info(&self) -> String {
+    fn debug_info(&self) -> String {
         // Unlike get_gd(), this doesn't require special trait bounds.
 
-        format!("{:?}", self.base)
+        format!("{:?}", self.base())
     }
 
     #[must_use]
-    pub fn into_raw(self) -> *mut Self {
+    fn into_raw(self) -> *mut Self
+    where
+        Self: Sized,
+    {
         Box::into_raw(Box::new(self))
     }
 
-    pub fn mark_destroyed_by_godot(&self) {
+    fn mark_destroyed_by_godot(&self) {
         out!(
             "    Storage::mark_destroyed_by_godot", // -- {:?}",
                                                     //self.user_instance
         );
-        self.lifecycle.set(Lifecycle::Destroying);
+        self.set_lifecycle(Lifecycle::Destroying);
         out!(
             "    mark;  self={:?}, val={:?}, obj={:?}",
             self as *const _,
-            self.lifecycle.get(),
-            self.base,
+            self.get_lifecycle(),
+            self.base(),
         );
     }
 
     #[inline(always)]
-    pub fn destroyed_by_godot(&self) -> bool {
+    fn destroyed_by_godot(&self) -> bool {
         out!(
             "    is_d;  self={:?}, val={:?}, obj={:?}",
             self as *const _,
-            self.lifecycle.get(),
-            self.base,
+            self.get_lifecycle(),
+            self.base(),
         );
-        matches!(self.lifecycle.get(), Lifecycle::Destroying)
+        matches!(self.get_lifecycle(), Lifecycle::Destroying)
     }
 }
+
+pub(crate) trait StorageRefCounted: Storage {
+    fn godot_ref_count(&self) -> u32;
+
+    fn on_inc_ref(&self);
+
+    fn on_dec_ref(&self);
+}
+
+#[cfg(not(feature = "experimental-threads"))]
+pub type InstanceStorage<T> = single_threaded::InstanceStorage<T>;
+#[cfg(feature = "experimental-threads")]
+pub type InstanceStorage<T> = multi_threaded::InstanceStorage<T>;
+
+pub type RefGuard<'a, T> = <InstanceStorage<T> as Storage>::RefGuard<'a>;
+pub type MutGuard<'a, T> = <InstanceStorage<T> as Storage>::MutGuard<'a>;
 
 impl<T: GodotClass> Drop for InstanceStorage<T> {
     fn drop(&mut self) {
         out!(
             "    Storage::drop (rc={})           <{:?}>",
             self.godot_ref_count(),
-            self.base,
+            self.base(),
         );
         //let _ = mem::take(&mut self.user_instance);
         //out!("    Storage::drop end              <{:?}>", self.base);
